LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/inline-subject-expected.txt

Hello world


FAIL View timeline attached to SVG graphics element assert_equals: expected "rgb(0, 0, 255)" but got "rgb(0, 128, 0)"
